ia-fastapi/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-multipart==0.0.12
pypdf==5.1.0
python-docx==1.1.2